requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
babel==2.18.0
backoff==2.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
courlan==1.3.2
dateparser==1.3.0
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
filetype==1.2.0
googleapis-common-protos==1.72.0
h11==0.16.0
haystack-ai==2.23.0
haystack-experimental==0.19.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
langfuse==3.13.0
langfuse-haystack==4.0.1
lazy_imports==1.2.0
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdit-plain==1.0.1
mdurl==0.1.2
more-itertools==10.8.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
ollama==0.6.1
ollama-haystack==6.0.0
openai==2.17.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pgvector==0.4.2
pgvector-haystack==6.1.0
posthog==7.8.3
protobuf==6.33.5
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tld==0.13.1
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
wrapt==1.17.3
zipp==3.23.0
